add working simulation
The simulation is now fonctional with torques only. It adds a KDL
chain for the simulated feedforward. The gains Kp and Kd are set
to an OK configuration that works with the damping set in the
URDF. If you modify the URDF damping to make the robot "slower"
you'll have to modify the gains again.